requirements.txt
Flask>=2.3.0
numpy>=1.25.0
Pillow>=10.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0 